Simplest connection FSM
I figure a gen_fsm is a good way to model a connection. This goes as
far as being able to connect to a server, but no further, and does not
handle errors very well at all.

I decided to provide good defaults for the frame records, for the time
being. Later I may put the defaults into the API rather than the
definitions. In any case, it'd be nice to not have to supply records
in the API.