Guaranteed delivery protocol for publishing
When publishing with `at_least_once` or `exactly_once` we record the
fact of the transmission; once the acknowledgement comes in, the
record is removed (or updated, in the case of `#pubrec{}`, which is an
intermediate step). I've not implemented retransmission yet.

(Also: I had 'at_least_once' and 'at_most_once' mixed up. OPS.)