Replay frames on connect if clean_session=false
The clean_session flag signifies connection stickiness; if it's
sticky, the server is expected to deliver messages that didn't get
acked the last time, and so are we.

This means I need to distinguish between RPCs and replayable
frames. The replayable frames are publish and pubrel; that is, those
sent by the publishing party.